Move exporter into agent-tooling
