--- agent/exporter/build.gradle
@@ -1,28 +1,0 @@
-plugins {
-  id "com.github.johnrengelman.shadow"
-}
-
-apply from: "$buildScriptsDir/common-java.gradle"
-
-repositories {
-  // maven local is needed for locally built otel fork
-  mavenLocal()
-}
-
-dependencies {
-  implementation "io.opentelemetry:opentelemetry-sdk:${versions.opentelemetry}"
-
-  // this is needed in order to access AiAppId for appId exchange data
-  compileOnly "io.opentelemetry.instrumentation:opentelemetry-instrumentation-api:${versions.opentelemetryInstrumentationAlpha}"
-
-  implementation "org.slf4j:slf4j-api:${versions.slf4j}"
-
-  implementation project(path: ":core")
-
-  implementation "com.azure:azure-monitor-opentelemetry-exporter:1.0.0-beta.4"
-  implementation "com.azure:azure-core:${versions.azureCore}"
-
-  testImplementation platform("org.junit:junit-bom:${versions.junit}")
-  testImplementation "org.junit.jupiter:junit-jupiter"
-  testImplementation "org.assertj:assertj-core:${versions.assertj}"
-}
