+++ settings.gradle
@@ -56,7 +56,6 @@ include ':agent:agent-gc-monitor:gc-monitor-api'
 include ':agent:agent-gc-monitor:gc-monitor-core'
 include ':agent:agent-gc-monitor:gc-monitor-tests'
 
-include ':agent:exporter'
 include ':agent:agent-profiler:agent-profiler-api'
 include ':agent:agent-profiler:agent-alerting-api'
 include ':agent:agent-profiler:agent-alerting'
